+++ config/index.js
@@ -18,9 +18,9 @@ const defaultRoot = resolve(process.env.HOME, ".cynomys", "protal");
 
 const [num, dataDir, logDir] = [cpus().length, resolve(defaultRoot, "data"), resolve(defaultRoot, "logs")];
 
-siteConfig.set("version", "v1.0.9bate"); // 设置服务器版本
+siteConfig.set("version", "1.0.10B"); // 设置服务器版本
 siteConfig.set("ssl", { // 安全设置
-  enable: true, // 是否开启ssl设置
+  enable: false, // 是否开启ssl设置
   keyName: "cynomy", // ssl公私钥 文件名 // 不写则默认
 // 存放位置,  path: "~/.ssh"  // 不写则默认位置
 });