+++ index.js
@@ -11,7 +11,8 @@
   */
 "use strict";
 
-const [{createServer}, {fork}, {resolve}, cluster] = [require("net"), require("child_process"), require("path"), require("cluster")];
+// const [{createServer}, {fork}, {resolve}, cluster] = [require("net"), require("child_process"), require("path"), require("cluster")];
+const [{createServer}, {fork}, {resolve}] = [require("net"), require("child_process"), require("path")];
 const {get} = require(resolve(__dirname, "config"));
 const [threadManager, slavePath ,bindHost, registryPort, {enable, slaveNum = 1}, remote] = [new Map(), resolve(__dirname, "distributed","slave"), get("bindHost"), get("registryPort"), get("distributed"), get("remote")];
 
@@ -23,6 +24,8 @@ const startSlave = socket => {
     const {act} = msg;
     if ("suicide" === act) {
       startSlave(socket);
+    } else if ("SHUTDOWN" === act) {
+      process.exit(0);
     }
   });
 
@@ -41,7 +44,7 @@ if (true === remote.enable) {
 
 if (enable) {
   console.log("portal server engine use cluster module, %s threads will startup", slaveNum);
-  cluster.schedulingPolicy = cluster.SCHED_RR; // 启用轮叫调度策略
+  // cluster.schedulingPolicy = cluster.SCHED_RR; // 启用轮叫调度策略 - 不启用轮叫调度模式
   const server = createServer();
   server.listen(registryPort, bindHost, () => {
     console.log("TCP net socket is runnig in %s:%s", bindHost, registryPort);